finished walking of account blob. Also created autopay type to use native mapping function between move and rust
