+++ ol/cli/src/node/query.rs
@@ -160,6 +160,75 @@ impl Node {
     }
 }
 
+
+
+
+/// check if the vec of value, is actually of other structs
+pub fn is_vec_of_struct(
+  move_val: &Vec<AnnotatedMoveValue>,
+) -> bool {
+    if let Some(e) = move_val.first() {
+      match e {
+        AnnotatedMoveValue::Struct(_) => return true,
+        _ => return false
+      }
+    }
+    false
+}
+
+/// get last vec
+pub fn get_last_stuct_in_vec(
+  move_val: &Vec<AnnotatedMoveValue>,
+) -> bool {
+    if let Some(e) = move_val.first() {
+      match e {
+        AnnotatedMoveValue::Struct(_) => return true,
+        _ => return false
+      }
+    }
+    false
+}
+
+// Ability to query Move types by walking an account blob. This is for structs which we may not have an equivalent type created in rust. For structs the core platform uses we have mappings available e.g. ol/types/src/miner_state.rs. This solves querying for resource structs that may be created by third parties.
+
+/// check if the vec of value, is actually of other structs
+pub fn unwrap_val_to_struct(
+  move_val: &AnnotatedMoveValue,
+) -> Option<&AnnotatedMoveStruct> {
+  match move_val {
+    AnnotatedMoveValue::Struct(s) => Some(s),
+    _ => None
+  }
+}
+
+/// find the value in a struct
+pub fn find_value_in_struct(s: &AnnotatedMoveStruct, key_name: String) -> Option<&AnnotatedMoveValue> {
+    match s
+        .value
+        .iter()
+        .find(|v| v.0.clone().into_string() == key_name)
+    {
+        Some((_, v)) => Some(v),
+        None => None,
+    }
+}
+/// finds a value
+pub fn find_value_from_state(
+    blob: &AnnotatedAccountStateBlob,
+    module_name: String,
+    struct_name: String,
+    key_name: String,
+) -> Option<&AnnotatedMoveValue> {
+    match blob.0.values().find(|&s| {
+        s.type_.module.as_ref().to_string() == module_name
+        && s.type_.name.as_ref().to_string() == struct_name
+    }) {
+        Some(s) => find_value_in_struct(s, key_name),
+        None => None,
+    }
+}
+
+
 /// test fixtures
 pub fn test_fixture_blob() -> AnnotatedAccountStateBlob {
     let mut s = BTreeMap::new();
@@ -184,32 +253,6 @@ pub fn test_fixture_struct() -> AnnotatedMoveStruct {
         is_resource: true,
         type_: module_tag.clone(),
         value: vec![(key, value)],
-    }
-}
-
-/// finds a value
-pub fn find_value_from_state(
-    blob: &AnnotatedAccountStateBlob,
-    module_name: String,
-    struct_name: String,
-    key_name: String,
-) -> Option<&AnnotatedMoveValue> {
-    match blob.0.values().find(|&s| {
-        // dbg!(&s.type_.name.as_ref().to_string());
-        s.type_.module.as_ref().to_string() == module_name
-            && s.type_.name.as_ref().to_string() == struct_name
-    }) {
-        Some(s) => {
-            match s
-                .value
-                .iter()
-                .find(|v| v.0.clone().into_string() == key_name)
-            {
-                Some((_, v)) => Some(v),
-                None => None,
-            }
-        }
-        None => None,
     }
 }
