+++ ol/onboard/src/manifest.rs
@@ -8,7 +8,7 @@ use ol_keys::scheme::KeyScheme;
 
 use libra_types::transaction::SignedTransaction;
 use libra_wallet::WalletLibrary;
-use ol_types::{account::ValConfigs, autopay::PayInstruction};
+use ol_types::{account::ValConfigs, pay_instruction::PayInstruction};
 use std::path::PathBuf;
 use crate::prelude::app_config;
