+++ ol/types/src/lib.rs
@@ -10,9 +10,10 @@
 )]
 
 pub mod account;
-pub mod autopay;
+pub mod pay_instruction;
 pub mod block;
 pub mod config;
 pub mod oracle_upgrade;
 pub mod miner_state;
 pub mod dialogue;
+pub mod autopay;
